Moved workflow to core
