+++ doc/source/api/force_bdss.core_plugins.dummy.dummy_kpi_calculator.tests.rst
@@ -4,6 +4,14 @@ force_bdss.core_plugins.dummy.dummy_kpi_calculator.tests package
 Submodules
 ----------
 
+force_bdss.core_plugins.dummy.dummy_kpi_calculator.tests.test_dummy_kpi_calculator module
+-----------------------------------------------------------------------------------------
+
+.. automodule:: force_bdss.core_plugins.dummy.dummy_kpi_calculator.tests.test_dummy_kpi_calculator
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 force_bdss.core_plugins.dummy.dummy_kpi_calculator.tests.test_dummy_kpi_calculator_bundle module
 ------------------------------------------------------------------------------------------------
