+++ doc/source/api/force_bdss.core_plugins.dummy.kpi_adder.tests.rst
@@ -12,6 +12,14 @@ force_bdss.core_plugins.dummy.kpi_adder.tests.test_kpi_adder_bundle module
     :undoc-members:
     :show-inheritance:
 
+force_bdss.core_plugins.dummy.kpi_adder.tests.test_kpi_adder_calculator module
+------------------------------------------------------------------------------
+
+.. automodule:: force_bdss.core_plugins.dummy.kpi_adder.tests.test_kpi_adder_calculator
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 
 Module contents
 ---------------
