--- doc/source/api/force_bdss.workspecs.rst
@@ -1,22 +1,0 @@
-force_bdss.workspecs package
-============================
-
-Submodules
-----------
-
-force_bdss.workspecs.workflow module
-------------------------------------
-
-.. automodule:: force_bdss.workspecs.workflow
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: force_bdss.workspecs
-    :members:
-    :undoc-members:
-    :show-inheritance:
